MOB-49 Content_items Key structure
